Neaten bootstrap script

+++ server_init.sh
@@ -2,19 +2,26 @@
 
 set -e
 
-cat /packages/keys/id_rsa.pub >> ~vagrant/.ssh/authorized_keys
+if [[ -f "~vagrant/bootstrapped" ]]; then
+  echo "Machine already bootstrapped, skipping..."
+  exit 0
+fi
 
-cd /packages
+# Add tribesports ubuntu repo to sources
+gpg --keyserver pgp.mit.edu --recv-keys D07E8C22
+gpg --export --armor D07E8C22 | apt-key add -
+source /etc/lsb-release
+cat > /etc/apt/sources.list.d/tribesports.list << EOF
+deb http://packages.tribesports.com/ubuntu ${DISTRIB_CODENAME}-tribesports main
+EOF
 
-apt-key add tribesports-pubkey.asc
-sudo -u vagrant -H gpg --import tribesports-privkey.asc || true
-
-echo 'deb http://packages.tribesports.com/ubuntu/ lucid-tribesports main' > /etc/apt/sources.list.d/tribesports.list
-
+# Upgrade the HECK out of things.
 apt-get update
 apt-get -y install aptitude
 # Conditional stuff here to handle 10.04/12.04
 DEBIAN_FRONTEND=noninteractive aptitude -y -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confold" full-upgrade
-apt-get install -y make s3cmd ts-ruby
+apt-get install -y s3cmd ruby1.9.3 dpkg-dev
 
 gem install fpm --no-ri --no-rdoc
+
+touch ~vagrant/bootstrapped
